final main

+++ src/java/model/JobPostings.java
@@ -1,17 +1,15 @@
 package model;
-
 import java.util.Date;
 import java.util.List;
 
 public class JobPostings {
-
     private int JobPostingID;
     private int RecruiterID;
     private String Title;
     private String Description;
     private String Requirements;
-    private double MinSalary;
-    private double MaxSalary;
+    private Long MinSalary;        // ✅ Thay double -> Long
+    private Long MaxSalary;        // ✅ Thay double -> Long
     private String Currency;
     private String Location;
     private Date PostedDate;
@@ -23,7 +21,7 @@ public class JobPostings {
     public JobPostings() {
     }
 
-    public JobPostings(int JobPostingID, int RecruiterID, String Title, String Description, String Requirements, double MinSalary, double MaxSalary, String Currency, String Location, Date PostedDate, Date ClosingDate, int Job_Posting_CategoryID, String Status) {
+    public JobPostings(int JobPostingID, int RecruiterID, String Title, String Description, String Requirements, Long MinSalary, Long MaxSalary, String Currency, String Location, Date PostedDate, Date ClosingDate, int Job_Posting_CategoryID, String Status) {
         this.JobPostingID = JobPostingID;
         this.RecruiterID = RecruiterID;
         this.Title = Title;
@@ -79,19 +77,19 @@ public class JobPostings {
         this.Requirements = Requirements;
     }
 
-    public double getMinSalary() {
+    public Long getMinSalary() {
         return MinSalary;
     }
 
-    public void setMinSalary(double MinSalary) {
+    public void setMinSalary(Long MinSalary) {
         this.MinSalary = MinSalary;
     }
 
-    public double getMaxSalary() {
+    public Long getMaxSalary() {
         return MaxSalary;
     }
 
-    public void setMaxSalary(double MaxSalary) {
+    public void setMaxSalary(Long MaxSalary) {
         this.MaxSalary = MaxSalary;
     }
